Restructure api.views/
